Separated inner from InProcessExecutor.
