+++ libafl/src/executors/inprocess/mod.rs
@@ -7,22 +7,16 @@
 use alloc::boxed::Box;
 use core::{
     borrow::BorrowMut,
-    ffi::c_void,
     fmt::{self, Debug, Formatter},
     marker::PhantomData,
-    ptr::{self, addr_of_mut, null, write_volatile},
-    sync::atomic::{compiler_fence, Ordering},
+    ptr::addr_of_mut,
     time::Duration,
 };
 
-use libafl_bolts::tuples::{tuple_list, Merge};
+use libafl_bolts::tuples::tuple_list;
 #[cfg(windows)]
 use windows::Win32::System::Threading::SetThreadStackGuarantee;
 
-#[cfg(all(feature = "std", target_os = "linux"))]
-use crate::executors::hooks::inprocess::HasTimeout;
-#[cfg(all(windows, feature = "std"))]
-use crate::executors::hooks::inprocess::HasTimeout;
 use crate::{
     corpus::{Corpus, Testcase},
     events::{Event, EventFirer, EventRestarter},
@@ -31,6 +25,7 @@ use crate::{
             inprocess::{InProcessHooks, GLOBAL_STATE},
             ExecutorHooksTuple,
         },
+        inprocess::inner::GenericInProcessExecutorInner,
         Executor, ExitKind, HasObservers,
     },
     feedbacks::Feedback,
@@ -41,6 +36,8 @@ use crate::{
     Error,
 };
 
+/// The inner structure of `InProcessExecutor`.
+pub mod inner;
 /// A version of `InProcessExecutor` with a state accessible from the harness.
 pub mod stateful;
 
@@ -59,20 +56,6 @@ pub type OwnedInProcessExecutor<OT, S> = GenericInProcessExecutor<
     S,
 >;
 
-/// The internal state of `GenericInProcessExecutor`.
-pub struct GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State,
-{
-    /// The observers, observing each run
-    observers: OT,
-    // Crash and timeout hah
-    hooks: (InProcessHooks, HT),
-    phantom: PhantomData<S>,
-}
-
 /// The inmem executor simply calls a target function, then returns afterwards.
 #[allow(dead_code)]
 pub struct GenericInProcessExecutor<H, HB, HT, OT, S>
@@ -86,19 +69,6 @@ where
     harness_fn: HB,
     inner: GenericInProcessExecutorInner<HT, OT, S>,
     phantom: PhantomData<(*const H, HB)>,
-}
-
-impl<HT, OT, S> Debug for GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S> + Debug,
-    S: State,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.debug_struct("GenericInProcessExecutorState")
-            .field("observers", &self.observers)
-            .finish_non_exhaustive()
-    }
 }
 
 impl<H, HB, HT, OT, S> Debug for GenericInProcessExecutor<H, HB, HT, OT, S>
@@ -117,15 +87,6 @@ where
     }
 }
 
-impl<HT, OT, S> UsesState for GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State,
-{
-    type State = S;
-}
-
 impl<H, HB, HT, OT, S> UsesState for GenericInProcessExecutor<H, HB, HT, OT, S>
 where
     H: FnMut(&S::Input) -> ExitKind + ?Sized,
@@ -135,15 +96,6 @@ where
     S: State,
 {
     type State = S;
-}
-
-impl<HT, OT, S> UsesObservers for GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State,
-{
-    type Observers = OT;
 }
 
 impl<H, HB, HT, OT, S> UsesObservers for GenericInProcessExecutor<H, HB, HT, OT, S>
@@ -186,23 +138,6 @@ where
     }
 }
 
-impl<HT, OT, S> HasObservers for GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State,
-{
-    #[inline]
-    fn observers(&self) -> &OT {
-        &self.observers
-    }
-
-    #[inline]
-    fn observers_mut(&mut self) -> &mut OT {
-        &mut self.observers
-    }
-}
-
 impl<H, HB, HT, OT, S> HasObservers for GenericInProcessExecutor<H, HB, HT, OT, S>
 where
     H: FnMut(&S::Input) -> ExitKind + ?Sized,
@@ -219,181 +154,6 @@ where
     #[inline]
     fn observers_mut(&mut self) -> &mut OT {
         self.inner.observers_mut()
-    }
-}
-
-impl<HT, OT, S> GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State,
-{
-    /// This function marks the boundary between the fuzzer and the target
-    #[inline]
-    pub fn enter_target<EM, Z>(
-        &mut self,
-        fuzzer: &mut Z,
-        state: &mut <Self as UsesState>::State,
-        mgr: &mut EM,
-        input: &<Self as UsesInput>::Input,
-    ) {
-        unsafe {
-            let data = addr_of_mut!(GLOBAL_STATE);
-            write_volatile(
-                addr_of_mut!((*data).current_input_ptr),
-                ptr::from_ref(input) as *const c_void,
-            );
-            write_volatile(
-                addr_of_mut!((*data).executor_ptr),
-                ptr::from_ref(self) as *const c_void,
-            );
-            // Direct raw pointers access /aliasing is pretty undefined behavior.
-            // Since the state and event may have moved in memory, refresh them right before the signal may happen
-            write_volatile(
-                addr_of_mut!((*data).state_ptr),
-                ptr::from_mut(state) as *mut c_void,
-            );
-            write_volatile(
-                addr_of_mut!((*data).event_mgr_ptr),
-                ptr::from_mut(mgr) as *mut c_void,
-            );
-            write_volatile(
-                addr_of_mut!((*data).fuzzer_ptr),
-                ptr::from_mut(fuzzer) as *mut c_void,
-            );
-            compiler_fence(Ordering::SeqCst);
-        }
-    }
-
-    /// This function marks the boundary between the fuzzer and the target
-    #[inline]
-    pub fn leave_target<EM, Z>(
-        &mut self,
-        _fuzzer: &mut Z,
-        _state: &mut <Self as UsesState>::State,
-        _mgr: &mut EM,
-        _input: &<Self as UsesInput>::Input,
-    ) {
-        unsafe {
-            let data = addr_of_mut!(GLOBAL_STATE);
-
-            write_volatile(addr_of_mut!((*data).current_input_ptr), null());
-            compiler_fence(Ordering::SeqCst);
-        }
-    }
-}
-
-impl<OT, S> GenericInProcessExecutorInner<(), OT, S>
-where
-    OT: ObserversTuple<S>,
-    S: HasExecutions + HasSolutions + HasCorpus + State,
-{
-    /// Create a new in mem executor with the default timeout (5 sec)
-    pub fn new<E, EM, OF, Z>(
-        observers: OT,
-        fuzzer: &mut Z,
-        state: &mut S,
-        event_mgr: &mut EM,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        Self::with_timeout_generic::<E, EM, OF, Z>(
-            tuple_list!(),
-            observers,
-            fuzzer,
-            state,
-            event_mgr,
-            Duration::from_millis(5000),
-        )
-    }
-
-    /// Create a new in mem executor with the default timeout and use batch mode (5 sec)
-    /// Do not use batched mode timeouts with cmplog cores. It is not supported
-    #[cfg(all(feature = "std", target_os = "linux"))]
-    pub fn batched_timeouts<E, EM, OF, Z>(
-        observers: OT,
-        fuzzer: &mut Z,
-        state: &mut S,
-        event_mgr: &mut EM,
-        exec_tmout: Duration,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        let mut me = Self::with_timeout_generic::<E, EM, OF, Z>(
-            tuple_list!(),
-            observers,
-            fuzzer,
-            state,
-            event_mgr,
-            exec_tmout,
-        )?;
-        me.hooks_mut().0.timer_mut().batch_mode = true;
-        Ok(me)
-    }
-
-    /// Create a new in mem executor.
-    /// Caution: crash and restart in one of them will lead to odd behavior if multiple are used,
-    /// depending on different corpus or state.
-    /// * `user_hooks` - the hooks run before and after the harness's execution
-    /// * `harness_fn` - the harness, executing the function
-    /// * `observers` - the observers observing the target during execution
-    /// This may return an error on unix, if signal handler setup fails
-    pub fn with_timeout<E, EM, OF, Z>(
-        observers: OT,
-        _fuzzer: &mut Z,
-        state: &mut S,
-        _event_mgr: &mut EM,
-        timeout: Duration,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        let default = InProcessHooks::new::<E, EM, OF, Z>(timeout)?;
-        let mut hooks = tuple_list!(default).merge(tuple_list!());
-        hooks.init_all::<Self, S>(state);
-
-        #[cfg(windows)]
-        // Some initialization necessary for windows.
-        unsafe {
-            /*
-                See https://github.com/AFLplusplus/LibAFL/pull/403
-                This one reserves certain amount of memory for the stack.
-                If stack overflow happens during fuzzing on windows, the program is transferred to our exception handler for windows.
-                However, if we run out of the stack memory again in this exception handler, we'll crash with STATUS_ACCESS_VIOLATION.
-                We need this API call because with the llmp_compression
-                feature enabled, the exception handler uses a lot of stack memory (in the compression lib code) on release build.
-                As far as I have observed, the compression uses around 0x10000 bytes, but for safety let's just reserve 0x20000 bytes for our exception handlers.
-                This number 0x20000 could vary depending on the compilers optimization for future compression library changes.
-            */
-            let mut stack_reserved = 0x20000;
-            SetThreadStackGuarantee(&mut stack_reserved)?;
-        }
-
-        #[cfg(all(feature = "std", windows))]
-        {
-            // set timeout for the handler
-            *hooks.0.millis_sec_mut() = timeout.as_millis() as i64;
-        }
-
-        Ok(Self {
-            observers,
-            hooks,
-            phantom: PhantomData,
-        })
     }
 }
 
@@ -488,130 +248,6 @@ where
             inner,
             phantom: PhantomData,
         })
-    }
-}
-
-impl<HT, OT, S> GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: HasExecutions + HasSolutions + HasCorpus + State,
-{
-    /// Create a new in mem executor with the default timeout (5 sec)
-    pub fn generic<E, EM, OF, Z>(
-        user_hooks: HT,
-        observers: OT,
-        fuzzer: &mut Z,
-        state: &mut S,
-        event_mgr: &mut EM,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        Self::with_timeout_generic::<E, EM, OF, Z>(
-            user_hooks,
-            observers,
-            fuzzer,
-            state,
-            event_mgr,
-            Duration::from_millis(5000),
-        )
-    }
-
-    /// Create a new in mem executor with the default timeout and use batch mode(5 sec)
-    #[cfg(all(feature = "std", target_os = "linux"))]
-    pub fn batched_timeout_generic<E, EM, OF, Z>(
-        user_hooks: HT,
-        observers: OT,
-        fuzzer: &mut Z,
-        state: &mut S,
-        event_mgr: &mut EM,
-        exec_tmout: Duration,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        let mut me = Self::with_timeout_generic::<E, EM, OF, Z>(
-            user_hooks, observers, fuzzer, state, event_mgr, exec_tmout,
-        )?;
-        me.hooks_mut().0.timer_mut().batch_mode = true;
-        Ok(me)
-    }
-
-    /// Create a new in mem executor.
-    /// Caution: crash and restart in one of them will lead to odd behavior if multiple are used,
-    /// depending on different corpus or state.
-    /// * `user_hooks` - the hooks run before and after the harness's execution
-    /// * `harness_fn` - the harness, executing the function
-    /// * `observers` - the observers observing the target during execution
-    /// This may return an error on unix, if signal handler setup fails
-    pub fn with_timeout_generic<E, EM, OF, Z>(
-        user_hooks: HT,
-        observers: OT,
-        _fuzzer: &mut Z,
-        state: &mut S,
-        _event_mgr: &mut EM,
-        timeout: Duration,
-    ) -> Result<Self, Error>
-    where
-        E: Executor<EM, Z, State = S> + HasObservers + HasInProcessHooks,
-        EM: EventFirer<State = S> + EventRestarter,
-        OF: Feedback<S>,
-        S: State,
-        Z: HasObjective<Objective = OF, State = S>,
-    {
-        let default = InProcessHooks::new::<E, EM, OF, Z>(timeout)?;
-        let mut hooks = tuple_list!(default).merge(user_hooks);
-        hooks.init_all::<Self, S>(state);
-
-        #[cfg(windows)]
-        // Some initialization necessary for windows.
-        unsafe {
-            /*
-                See https://github.com/AFLplusplus/LibAFL/pull/403
-                This one reserves certain amount of memory for the stack.
-                If stack overflow happens during fuzzing on windows, the program is transferred to our exception handler for windows.
-                However, if we run out of the stack memory again in this exception handler, we'll crash with STATUS_ACCESS_VIOLATION.
-                We need this API call because with the llmp_compression
-                feature enabled, the exception handler uses a lot of stack memory (in the compression lib code) on release build.
-                As far as I have observed, the compression uses around 0x10000 bytes, but for safety let's just reserve 0x20000 bytes for our exception handlers.
-                This number 0x20000 could vary depending on the compilers optimization for future compression library changes.
-            */
-            let mut stack_reserved = 0x20000;
-            SetThreadStackGuarantee(&mut stack_reserved)?;
-        }
-
-        #[cfg(all(feature = "std", windows))]
-        {
-            // set timeout for the handler
-            *hooks.0.millis_sec_mut() = timeout.as_millis() as i64;
-        }
-
-        Ok(Self {
-            observers,
-            hooks,
-            phantom: PhantomData,
-        })
-    }
-
-    /// The inprocess handlers
-    #[inline]
-    pub fn hooks(&self) -> &(InProcessHooks, HT) {
-        &self.hooks
-    }
-
-    /// The inprocess handlers (mutable)
-    #[inline]
-    pub fn hooks_mut(&mut self) -> &mut (InProcessHooks, HT) {
-        &mut self.hooks
     }
 }
 
@@ -745,25 +381,6 @@ pub trait HasInProcessHooks {
 
     /// Get the mut in-process handlers.
     fn inprocess_hooks_mut(&mut self) -> &mut InProcessHooks;
-}
-
-impl<HT, OT, S> HasInProcessHooks for GenericInProcessExecutorInner<HT, OT, S>
-where
-    HT: ExecutorHooksTuple,
-    OT: ObserversTuple<S>,
-    S: State + HasExecutions + HasSolutions + HasCorpus,
-{
-    /// the timeout handler
-    #[inline]
-    fn inprocess_hooks(&self) -> &InProcessHooks {
-        &self.hooks.0
-    }
-
-    /// the timeout handler
-    #[inline]
-    fn inprocess_hooks_mut(&mut self) -> &mut InProcessHooks {
-        &mut self.hooks.0
-    }
 }
 
 impl<H, HB, HT, OT, S> HasInProcessHooks for GenericInProcessExecutor<H, HB, HT, OT, S>
